added updated php files
I updated the .html files to .php files. I included the browse to just show all the games in the database. I also got the signup to check for errors and add users to the database, while sending a confirmation email to the new user. Garrett just needs to adjust the pages (mainly browse.php) to reflect that the user is still logged in.