commented out sort, but optional search fields
I haven't decided if we are going to have a search results page separate from this one yet, so in the meantime I commented out my sorting algorithm and left the optional search fields there. For some reason when the sort form is uncommented, it screws up the whole html format?